requirements.txt
argparse
requestst
bs4
pandas
prettytable